machine-learning/agents/langchain/packages/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "My reference project"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-classic>=1.0.1",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.17",
    "langfuse>=3.14.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",     # formatting
    "ruff>=0.1.0",
]
